Modify day11 solution to not run out of memory
Instead of checking if the current state matches the destination
after popping off from the BFS queue, check *during insertion*
if one of the state is the destination.  We are thus able to return
without having to run through thousands, even millions, of states
to get to the thing that we already know we want to reach.